Warn about implicit conversions that result in precision loss (#169)
* Warn about implicit conversions that result in precision loss

MSVC warns about those anyway with /W4. Having them pop up in POSIX
environments saves us finding out about the warning from CI.

* Fix 32-bit warnings

+++ source/encoding.c
@@ -184,7 +184,7 @@ int aws_hex_decode(const struct aws_byte_buf *AWS_RESTRICT to_decode, struct aws
             return aws_raise_error(AWS_ERROR_INVALID_HEX_STR);
         }
 
-        uint8_t value = high_value << 4;
+        uint8_t value = (uint8_t)(high_value << 4);
         value |= low_value;
         output->buffer[written++] = value;
     }
